+++ core/src/main/java/com/hlag/oversigt/util/text/JsonPathFunction.java
@@ -4,18 +4,16 @@ import java.util.Collection;
 import java.util.Objects;
 import java.util.function.Function;
 
-import com.jayway.jsonpath.Configuration;
+import com.hlag.oversigt.util.JsonUtils;
 import com.jayway.jsonpath.JsonPath;
 
 import edu.umd.cs.findbugs.annotations.Nullable;
 
 class JsonPathFunction implements Function<String, String> {
-	private final Configuration configuration;
 
 	private final String json;
 
-	JsonPathFunction(final Configuration configuration, final String json) {
-		this.configuration = configuration;
+	JsonPathFunction(final String json) {
 		this.json = json;
 	}
 
@@ -23,7 +21,7 @@ class JsonPathFunction implements Function<String, String> {
 	public String apply(@Nullable final String jsonPathString) {
 		Objects.requireNonNull(jsonPathString, "Input for a JsonPath must be non null");
 		final JsonPath jsonPath = JsonPath.compile(jsonPathString);
-		final Object result = jsonPath.read(json, configuration);
+		final Object result = JsonUtils.extractValueUsingJsonPath(jsonPath, json);
 
 		if (result == null) {
 			return "";
